# docs/swagger.yaml
basePath: /
definitions:
...
  quarter_api.TrimestreGetResponse:
    properties:
      ano:
        type: integer
      codigo:
        type: string
      data_fim:
        type: string
      data_inicio:
        type: string
      id:
        type: integer
      trimestre:
        type: integer
    type: object
info:
  contact:
    email: crisaltmann@gmail.com
    name: Cristiano Altmann
  description: Swagger API for Fundament Stock Api.
  license:
    name: MIT
    url: https://github.com/crisaltmann/fundament-stock-api
  termsOfService: http://swagger.io/terms/
  title: Fundament Stock Api Swagger API
  version: "1.0"
paths:
...
  /quarters/{id}:
    get:
      parameters:
      - description: quarter id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quarter_api.TrimestreGetResponse'
      summary: Retorna o trimestre
swagger: "2.0"